Narrow Bo storage from f64 to f32 in MarsData
- kernels.{h,cc}: orthonormalize()'s Bo parameter goes from
  const double* to const float*. Split the inner kernels by operand
  type: axpy_m and dot_bo take f32 bo_row (cvtps_pd at the load);
  dot_m is now restricted to the f64*f64 case (T dot T for the DGKS
  gate, the only remaining f64*f64 use). compute_BoT_bx_col /
  project_subtract_and_norm take const float* Bo.

- marsalgo.cc: MarsData::Bo is MatrixXfC. init writes
  sqrt_w.cast<float>(); ybo computes via .cast<double>().transpose().
  append()'s Gram-Schmidt and DGKS retry lazily cast each Bo column to
  f64 (Eigen expression, no temp) so the projection arithmetic stays
  f64. The final v/w is downcast to f32 on the store back to
  Bo.col(_m). Bok rebuild loop drops its .cast<float>() (both sides
  are now f32). Per-thread scratch doc updated.

- tests/test_kernels.cc: typedef MatrixXfC.
  make_orthonormal_basis returns MatrixXfC (builds f64 internally,
  casts at the end -- matches what append() does after Gram-Schmidt).
  eigen_reference takes MatrixXfC Bo, casts to f64 once for the
  matmul. Tests cast Bo to MatrixXd as needed for f64 assertions, and
  the (Bo^T Bo) ~= I sanity check relaxes from TOL=1e-14 to F32_TOL.

Tests: 13/13 C++ + 20/20 pytest pass with no further tolerance
changes (the f32 Bx tolerances already absorbed the Bo narrowing
since both end up in the same f32 storage floor).

Memory: Bo footprint halves (n * max_terms * 8 -> n * max_terms * 4),
single instance per fit. At n=millions, M=400 that's ~5 GB saved.
Per-thread savings still come from the prior Bx narrowing.

Bok drop is the natural follow-up: now that Bo is f32 with the same
row width as Bok, on-the-fly gather (with software prefetch) can
replace the pre-sorted scratch copy.

+++ kernels.h
@@ -24,17 +24,17 @@ constexpr double DGKS_GATE_RATIO_SQ = 9.0;
  *      s        = ||Bx[:,j]||^2                            // f64 accumulator over post-store f32
  *      Bx[:,j] *= (s > tol) ? 1/sqrt(s + tol) : 0          // normalize, zero degenerate
  *
- *  Bx is stored as f32 to halve the per-thread scratch footprint; all
- *  arithmetic (projection, dot products, norms) stays in f64. The
- *  cancellation in (Bx -= Bo * Bo^T * Bx) is bounded by the DGKS retry, so
- *  the rounded f32 storage keeps O(eps_f32) orthogonality against Bo --
- *  see DGKS_GATE_RATIO_SQ below.
+ *  Bo and Bx are both stored as f32 to halve the largest buffers; all
+ *  arithmetic (projection, dot products, norms) stays in f64 -- Bo and Bx
+ *  entries are upcast at the load. The cancellation in (Bx -= Bo*Bo^T*Bx)
+ *  is bounded by the DGKS retry, so the rounded f32 storage keeps
+ *  O(eps_f32) orthogonality against Bo -- see DGKS_GATE_RATIO_SQ below.
  *
  *  Layouts (caller's responsibility):
  *      B     : (n, *) col-major float;  col stride ldB.   Only columns mask[j] are read.
  *      x     : (n)    float
  *      mask  : (p)    int32, indexes into columns of B
- *      Bo    : (n, m) row-major double; row stride ldBo
+ *      Bo    : (n, m) row-major float;  row stride ldBo
  *      Bx    : (n, p) col-major float;  col stride ldBx       [output]
  *      T     : (m, p) col-major double; col stride ldT        [workspace; overwritten]
  *      s_buf : (p)    double                                  [workspace; overwritten]
@@ -53,7 +53,7 @@ void orthonormalize(
     const float  *B,    int ldB,
     const float  *x,
     const int    *mask,
-    const double *Bo,   int ldBo,
+    const float  *Bo,   int ldBo,
     float        *Bx,   int ldBx,
     double       *T,    int ldT,
     double       *s_buf,
